+++ src/Web/Sfa.Das.Sas.Web/ViewModels/ProviderFrameworkSearchResultViewModel.cs
@@ -8,7 +8,7 @@ namespace Sfa.Das.Sas.Web.ViewModels
     {
         public string Title { get; set; }
 
-        public double TotalResults { get; set; }
+        public long TotalResults { get; set; }
 
         public int ResultsToTake { get; set; }
